chore: move example code to top-level
