--- ios/example/data/a.txt
@@ -1,1 +1,0 @@
-This is the a.txt file! Exciting, isn't it?
