--- ios/example/data/b.txt
@@ -1,1 +1,0 @@
-You may have already guessed, but this is the b.txt file. Excruciating to look upon, isn't it?
